Move custom ServerConfig c'tor to mod config
And also move impl TryFrom<&AccountConfig> for ServerAccountConfig

Addresses review comments https://github.com/NomicFoundation/hardhat/pull/4221#discussion_r1284958628
and https://github.com/NomicFoundation/hardhat/pull/4221#discussion_r1281937751
